flask-api/requirements.txt
Flask==2.0.1
fluent-logger==0.10.0